+++ src/main/java/sa/gov/nic/bio/bw/client/features/registerconvictedpresent/webservice/ConvictedReport.java
@@ -1,5 +1,6 @@
 package sa.gov.nic.bio.bw.client.features.registerconvictedpresent.webservice;
 
+import sa.gov.nic.bio.bw.client.features.commons.webservice.Finger;
 import sa.gov.nic.bio.bw.client.features.commons.webservice.Name;
 
 import java.util.List;